introduce new KVPool to aid in concurrent usage
The KVPool addresses the use case where an application
wants to support N concurrent actors on a single
ForestDB KVStore at one time.  ForestDB requries that
these actors operate on separate fdb handles and this
KVPool attempts to make this pattern simpler.

Users invoke NewKVPool, with parameters used
to create the *File and child *KVStore objects as well
as the number N of objects to pool.  Only references
to the *KVStore are kept in the pool, access to the
parent *File is done through the KVStore.File()
method.

Once you have a KVPool, clients get a KVStore by using
the Get() method.  The MUST return it using the
Return() method when done.  The client MUST call
Close() in order to free all resources.  The call
to Close() may block forever in the event that
KVStores acquired through Get() were not Return()'d.
Change-Id: Icacc4f3fd0daf7957bb82c5a6f4b1facf3461acc
Reviewed-on: http://review.couchbase.org/55706
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Marty Schoch <marty.schoch@gmail.com>